--- a/app/src/app/store/unassignedFeatures.ts
+++ b/app/src/app/store/unassignedFeatures.ts
@@ -1,5 +1,4 @@
 import {useChartStore} from '@/app/store/chartStore';
-import {idCache} from '@/app/store/idCache';
 import {useMapStore} from '@/app/store/mapStore';
 import GeometryWorker from '@/app/utils/GeometryWorker';
 import {LngLatBoundsLike} from 'maplibre-gl';
@@ -42,8 +41,8 @@ export const useUnassignFeaturesStore = create<UnassignedFeatureStore>((set,get)
     const {shatterIds, zoneAssignments, mapDocument, getMapRef} = useMapStore.getState();
     const mapRef = getMapRef();
     if (!GeometryWorker || !mapRef) return;
-    const useBackend =
-      idCache.getTotalPopSeen(shatterIds.parents) !== useChartStore.getState().chartInfo.totPop;
+    const useBackend = true
+      // idCache.getTotalPopSeen(shatterIds.parents) !== useChartStore.getState().chartInfo.totPop;
     if (!useBackend) {
       await GeometryWorker.updateProps(Array.from(zoneAssignments.entries()));
     }