--- a/app/src/app/utils/GeometryWorker/geometryWorker.ts
+++ b/app/src/app/utils/GeometryWorker/geometryWorker.ts
@@ -45,20 +45,20 @@ const GeometryWorker: GeometryWorkerClass = {
     this.geometries = {};
   },
   loadGeometry(featuresOrStringified, idProp) {
-    // const features: MapGeoJSONFeature[] =
-    //   typeof featuresOrStringified === 'string'
-    //     ? JSON.parse(featuresOrStringified)
-    //     : featuresOrStringified;
-    // const firstEntry = Object.values(this.geometries)[0];
-    // features.forEach(f => {
-    //   const id = f.properties?.[idProp];
-    //   // TODO: Sometimes, geometries are split across tiles or reloaded at more detailed zoom levels
-    //   // disambiguating and combining them could be very cool, but is tricky with lots of edge cases
-    //   // and computationally expensive. For now, we just take the first geometry of a given ID
-    //   if (id && !this.geometries[id]) {
-    //     this.geometries[id] = structuredClone(f);
-    //   }
-    // });
+    const features: MapGeoJSONFeature[] =
+      typeof featuresOrStringified === 'string'
+        ? JSON.parse(featuresOrStringified)
+        : featuresOrStringified;
+    const firstEntry = Object.values(this.geometries)[0];
+    features.forEach(f => {
+      const id = f.properties?.[idProp];
+      // TODO: Sometimes, geometries are split across tiles or reloaded at more detailed zoom levels
+      // disambiguating and combining them could be very cool, but is tricky with lots of edge cases
+      // and computationally expensive. For now, we just take the first geometry of a given ID
+      if (id && !this.geometries[id]) {
+        this.geometries[id] = structuredClone(f);
+      }
+    });
   },
   loadRectFeatures(featureDict) {
     Object.entries(featureDict).forEach(([id, feature]) => {