--- a/app/src/app/utils/helpers.ts
+++ b/app/src/app/utils/helpers.ts
@@ -76,14 +76,13 @@ export const boxAroundPoint = (
 let USE_RTREE = true
 let searchTimesPer1kFeatures: number[] = []
 
-setInterval(() => {
-  console.log('Time in MS to query 1000 features:', searchTimesPer1kFeatures.reduce((a,b)=>a+b,0)/searchTimesPer1kFeatures.length)
-}, 1000);
-
 export const toggleUseRTree = () => {
   USE_RTREE = !USE_RTREE
   searchTimesPer1kFeatures = []
   alert(USE_RTREE ? "Using RTree" : "Not using RTree")
+}
+export const getSearchTimes = () => {
+  alert(`Time in MS to query 1000 features: ${searchTimesPer1kFeatures.reduce((a,b)=>a+b,0)/searchTimesPer1kFeatures.length}`)
 }
 /**
  * getFeaturesInBbox
@@ -112,10 +111,7 @@ export const getFeaturesInBbox = (
       maxX: bottomRightLatLon.lng,
       minY: bottomRightLatLon.lat
     }
-    const t0 = performance.now();
     const features = featureCache.searchFeaturesinBbox(bboxLatLon)
-    const t1 = performance.now();
-    searchTimesPer1kFeatures.push((t1 - t0)/features.length*1000)
     return filterFeatures(features as any, filterLocked);
   } else {
     const bbox = boxAroundPoint(e, brushSize);